// planning-idsdp-pps-frontend/src/app/main/modules/rdpp-rtapp-management/features/dpp-tapp-list/project-details/models/other-details.model.ts
import {ProjectDetailsPartB} from "./project-details-partb.model";

export class OtherDetails {
    uuid: string;
    projectConceptUuid: string;
    specificationLinkagePerspective: string;
    contributionProjectAchieving: string;
    relationProjectAllocation: string;
    whetherPrivateSector: string;
    majorConditionalityForeignAid: string;
    involvementCompensation: string;
    riskAnalysisMitigation: string;
    projectDetailsPartBRequest: ProjectDetailsPartB;
}